remove DSL_DataSpace_LegendQueryApplicationPlugin
